permit bill-details

+++ backend/src/routes/private/bill-details.js
@@ -1,4 +1,3 @@
-/* eslint-disable require-atomic-updates */
 const router = require('express').Router();
 const knex = require('../../knex');
 const { validateCustomer } = require('../../middlewares/middlewarAuthorize');
@@ -6,12 +5,18 @@ const { handleAPIResponse } = require('../../common/handleAPIResponse');
 
 router.route('/api/restaurants/:restaurant_id/bills/:bill_id')
   .get(validateCustomer, async (req, res, next) => {
+    const { user_id, position } = req.session;
     const { restaurant_id, bill_id } = req.params;
     if (!restaurant_id || !bill_id) return handleAPIResponse(res, 400, 'restaurant_id && bill_id required');
     try {
-      //
-      const bill = await knex('bills').first().where({ restaurant_id, id: bill_id });
+      const [restaurant, bill] = await Promise.all([
+        knex('restaurants').first().where({ id: restaurant_id }),
+        knex('bills').first().where({ restaurant_id, id: bill_id }),
+      ]);
+      if (restaurant.manager_id !== user_id && bill.recipient_id !== user_id && position !== 'admin') return handleAPIResponse(res, 403, 'forbidden');
+      if (!restaurant) return handleAPIResponse(res, 404, 'restaurant_id not exist');
       if (!bill) return handleAPIResponse(res, 404, 'bill_id not exist');
+
       const bill_detail = await knex('bill_detail').where({ bill_id: bill.id });
 
       return handleAPIResponse(res, 200, { ...bill, bill_detail });
@@ -67,12 +72,17 @@ router.route('/api/restaurants/:restaurant_id/bills/:bill_id')
   //   }
   // })
   .delete(validateCustomer, async (req, res, next) => {
+    const { user_id, position } = req.session;
     const { restaurant_id, bill_id } = req.params;
     if (!restaurant_id || !bill_id) return handleAPIResponse(res, 400, 'restaurant_id && bill_id required');
 
     try {
-      //
-      const bill = await knex('bills').first().where({ restaurant_id, id: bill_id });
+      const [restaurant, bill] = await Promise.all([
+        knex('restaurants').first().where({ id: restaurant_id }),
+        knex('bills').first().where({ restaurant_id, id: bill_id }),
+      ]);
+      if (restaurant.manager_id !== user_id && bill.recipient_id !== user_id && position !== 'admin') return handleAPIResponse(res, 403, 'forbidden');
+      if (!restaurant) return handleAPIResponse(res, 404, 'restaurant_id not exist');
       if (!bill) return handleAPIResponse(res, 404, 'bill_id not exist');
       await Promise.all([
         knex('bills').delete().where({ restaurant_id, id: bill_id }),
